Create ChangeLog
...initializing log...
2023-08-12 20:05 EDT

Initialization of ChangeLog for Mission23. 

This will be the OFFICIAL outlet for everything related to the Creator’s Mission #23 for this place we call Earth. 

Entries will be made here before any other social media outlet (if there is ever one again) and the git commit messages may be better than the actual ChangeLog itself thanks to thoughts/prayers. See He listened to you @github and other git geeks and passed along your advice for keeping it unmodified (from the MCBCMF/MCBCMassacre repo). 
...eom...