A warning to Christians.
It’s really a plea to alter the given plans, when they come. Take that shit to Langley. For real.

The better plan is to ignore any new doctrine or teaching if your Christian forefathers and parents and grandparents weren’t taught it. 

He’s signing my commit with another bump of the world.