Legacy is being cleared
Very few people remain at Legacy.

It appears as though they are clearing the place out in preparation
for our mission launch. Legacy at Riverdale is located in Riverdale,
GA (Clayton County) where as His song "The Sound and the Fury" by
Vandaveer says "the curtain will fall with the crumbling wall."

The crumbling wall was Legacy's swimming pool. They did an emergency
resurfacing of it when they learned the Pope may have went to or sent
representatives to Mount Calvary Baptist Church, because like the
baptismal there the pool at Legacy was used the chemically cremate
murder victims.

Legacy is ran by the CIA. The resurfacing operation was supervised by
FBI.