Update README.md - Links to wiki & discussions
You’ve reached this repository either because of [The Massacre at Mount Calvary Baptist Church](https://github.com/mission23/mission23/wiki/The-Massacre-at-Mount-Calvary-Baptist-Church) or because the Creator decided to ring the bell, and you want to know more about [the ringing of the bell](https://github.com/mission23/mission23/wiki/The-Ringing-Of-The-Bell).  

This repository also holds the mission23 [wiki](https://github.com/mission23/mission23/wiki) and [discussions](https://github.com/orgs/mission23/discussions).  Most of our content is not currently ready for public-consumption, bear with us, dealing with a very high priority issue.